Android: don't show the zoom controls by default.
We don't have any APIs for toggling the visibility of the zoom controls
in Qt and the support on different platforms might vary, so we should
just hide the controls by default. On devices with API level less then
11 the controls will still be visible, as there is no way to disable
them. Note that this does _not_ mean we are disable zooming.

Change-Id: Ie49c4bc4e8441b6b1bdc0a0db0d728d8934a14eb
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@theqtcompany.com>