iOS: transfer focus to QWindowControllerItem on touch release
When the user touches the native web view, we need to inform QtQuick
that the window controller should now have focus so that we
clear focus from other items in the scene. Otherwise HTML items
inside the webview will have focus (with blinking cursors etc) at
the same time as an unrelated QtQuick item.

We achieve this by installing a gesture recogniser on
the web view. The gesture catches touch release done on the webview
and calls setFocus on the wrapper item.

Change-Id: I69f21fc226eb48c5640f9ae473562c56baf81e4a
Reviewed-by: Morten Johan Sørvig <morten.sorvig@digia.com>