Refactor QWebView code.
In an attempt to make the code more future proof, this patch tries to
improve the abstraction between the native WebViews/View controllers
and the public APIs. This should make it easier to add new platforms
and alternative public interfaces, e.g., a Window/Widget API. It should
also make it easier to implement a plugin system if needed.

In addition the following changes were done:
- reload() implementation in the android controller.
- reload() calls the native reload function on both Android and iOS.
- loadProgress() calls the native getProgress() directly on Android.
- isLoading() calls loading() on the UIWebView (iOS)
- runJavaScript() will now only pass the script and an id to the platform
  implementation to avoid creating a strong coupling to QJSValue.

Change-Id: I3cbd81c7fd8d93bacf9134be32ad061d1f9e1183
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@theqtcompany.com>
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@theqtcompany.com>