Android: Fix isLoading() and progress tracking.
There are no good ways to know if a WebView is loading or not.
onPageStarted() is called for each frame, so it might be called
multiple times while onPageFinished() is only called for the main
frame, that is, resources and other frames might still be loading.
To make it worse load status is handled in a different client then
the progress, so to make sure we report progress and finished in
the right order we need to sync them up and let the last call
be responsible for reporting the loading succeeded status.

Change-Id: Ibcf858bfa026c5b280c70b338801be9a1475b376
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@theqtcompany.com>