Return progress, title and url from the cached values in QWebView.
Returning the cached values for progress, title and url might be more
accurate in some cases, e.g., on Android we might not get the correct
url if we query the native WebView immediately after receiving the
url-changed callback. Using the already cached values also removes the
need to cache them again in the platform implementation.

Change-Id: I3a9d362fe5e5a47a823da2d84fd638c409b1bcf3
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@theqtcompany.com>
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@theqtcompany.com>