Fix the geometry source in QQuickViewController.
The controller item was using the rect it got from geometryChanged()
and mapped it to scene coordinates, but then in its own coordinate
system. The problem with this is that the rect we get, in
geometryChanged(), is relative to the view-controllers parent and can
therefore carry an offset that will give us the wrong scene coordinates.
The easiest solution is to remove the offset and place the rect at
(0, 0), e.g., by calling clipRect().

Change-Id: I29bf729e82498a2534425dc4c173f055305a1b2d
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@theqtcompany.com>
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@theqtcompany.com>