Android: Fix setting geometry and clip rect.
All native views are considered top-level, that is, the view expects its
geometry to be in the global coordinate system and therefore clipping
and mapping from scene to global coordinates needs to be done manually.
Previously some of this was handle in the foreign platform window
implementation, but since there are no real child/parent relationship
in the platform implementation, it made it impossible to handle the case
where the scene was not covering the whole screen and/or if the scene
could move or be re-sized (e.g., QQuickWidget).

Task-number: QTBUG-43391
Change-Id: I019f76b6cec1841a480f901e14432800cb3da817
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@theqtcompany.com>