Android: Add protection against deadlocking and/or triggering ANR.
We block the Android UI thread when we call into the native WebView
which means it's possible to deadlock or cause the
"Application Not Responding" handler to be activated, e.g., if the
WebView or UI thread is unable to process the request quickly enough.
This change makes sure we never block the UI thread for longer then
250ms. Anything beyond 200ms is and we can be certain the UI thread is
overloaded or we're blocking both the Qt thread and the Android thread.

Change-Id: I237674091e4d5db80209b5295f75b61e886f3db4
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@theqtcompany.com>