Add initialize function
The initialization function will set-up any platform specific
requirements needed to display the WebView correctly.
Initially we tried to avoid this, as it's not a very nice approach, and
it was only needed when QtWebEngine back-end was the used (see:
QtWebEngine::initialize()). Since we now require special set-up for OS X
as well, it has become unreasonable to expect that the user should
implement this logic in their application and that it's more convenient
if we provide a uniform solution to make the WebView work on all
platforms.
The implication of this change, is that we now actually expose a C++ API
and will therefore need to export the module again
(making it a non-internal module).

Change-Id: I87678db7cd50b25465bcf623c008f04a4cffd7df
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@theqtcompany.com>