Android: Explicitly tear-down the Native WebView.
The WebView will continue to evaluate JavaScript even after we've
released all handles to the WebView, to avoid this we'll call destroy()
when we release the WebView on our side. On Android 6.0 there's issued
a warning when destroy is called, but this is not fatal and seems to
be a know issue on their side. Alternatively, if it turns out that
calling destroy() causes problems in the future, it's possible to stop
JS evaluation by disabling it through the WebSettings interface.

Task-number: QTBUG-49664
Change-Id: Ic98d24d58525397ad06ecbc4476cd5613b6663ab
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@theqtcompany.com>