Add a proper backend for WebEngine
This removes the need for the awkward type re-registering and renaming
that was done in the Qml plugin, to support QtWebEngine.
Among other issues, the main motivation for this change is to unify the
implementations, and make it possible to run the auto tests that
interfaces directly with the QQuickWebView class. Having an actual
backend for Qt WebEngine means we can get rid of the, non-functional,
default implementation as well.

Change-Id: Ia93611ed2755c92207ca86ba3890ac4c2c4d72e9
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@theqtcompany.com>