Make it possible to put a WebView inside a QQuickWidget on iOS
When setting the parent view of the WebvView we need to use the real
window, and not the off-screen window provided by QQuickWidget.
This change also improves the logic for setting the geometry, as extra
care is needed to make sure the view is position correctly inside the
scene when using a QQuickWidget.

Task-number: QTBUG-48221
Change-Id: I476cac926df6c862cc3602b21e97003cc254726c
Reviewed-by: Timur Pocheptsov <timur.pocheptsov@theqtcompany.com>