Android: Don't call setClipBounds on the WebView
When High DPI mode is enabled, with a scale different then 1, the values
we get are not in the same scale as setClipBounds expects. While it is
possible to convert the values to native pixel first, it's probably
better if we just disable this functionality on Android, that is, until
a better cross-platform solution can be found.

Task-number: QTBUG-51198
Change-Id: Ia3e3e3911b9ba24d83e49afa6bec61696b93073a
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@theqtcompany.com>