Fix macOS 10.9 builds that use macOS 10.10 SDK
WebFrameLoadDelegate was an informal protocol in <= 10.10 SDK, which
means it should not be specified in the protocol conformance list.
Doing so would cause an error that the specified protocol declaration
can not be found.

Starting with macOS 10.11 SDK, it is a formal protocol, and can thus
be used in the conformance list.

Change-Id: I6bb92e202735c7c00dd19eb2b858e7fd327be89e
Task-number: QTBUG-56535
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>