qquickwebviewcontroller: don't remove listener from wrong parent
QQuickViewChangeListener::itemParentChanged is called for parent
changes done to any of the items we listen to up the parent chain.
But as it stood, we would always remove the change listener from
m_item->parentItem(), even if the changed item was higher up in
the hierarchy. We would therefore wrongly miss any later changes
done to m_item->parentItem().

A bug from this could be seen when pushing an item containing a
webview to a stackview, since then, several parent changes would
occur.

[ChangeLog][General] Fixed missing geometry update bug when a
webview changed anchestor (e.g. when pushing it onto a StackView).

Task-number: QTBUG-54128
Change-Id: Iff00a0029dd713306db7e39db8c92672a6b379c7
Reviewed-by: Christian Stromme <christian.stromme@qt.io>