Make QtWebView plugin based
This removes the hard build dependency to QtWebEngine, which opens up
the possibility for QtWebEngine, or others, to provide their own plugin.
Another benefit of having the backends loaded at run-time, is that we
can provide an alternative for developers that wants to publish their
application in the App Store, where shipping QtWebEngine isn't an option,
due to store policies, and where we already have an alternative/experimental
backend that can be used.

[ChangeLog][WebView] QtWebView will now load its backends at run-time.

Task-number: QTBUG-63137
Change-Id: I581940fe4c3b5e6bb41896367d3163ac8bc7b6b9
Reviewed-by: Kai Koehne <kai.koehne@qt.io>