WKWebView: try to pass URLs with unrecognized schemes down to the OS
This more closely matches the behavior of the Android implementation,
which attempts to launch an Intent to handle URLs with schemes not
directly handled by the web view itself. The QtWebEngine implementation
similarly passes URLs with unhandled schemes down to QDesktopServices.
For the native web view API we'll go directly to the system URL handling
APIs instead of through QDesktopServices again in order to match the
Android implementation which uses the native API (Intents).

Change-Id: Ie2881bf17f3b057cc053c9a4faa8cd81d6f6faca
Reviewed-by: Jake Petroules <jake.petroules@qt.io>