winrt: Handle setting of URL with unsupported scheme
While the webview handles clicks on links with unsupported schemes (like
mailto or sms) internally and opens the default program/asks the user
what to do, calling the url directly resulted in an assert.
We try to mimic the webview's behavior as closely as possible by
launching the default program for the given URI. If no default program
is set, the user is asked, what to do.

Change-Id: I4b5220f3de76ebecb50643110944056a60c3b3b6
Reviewed-by: Maurice Kalinowski <maurice.kalinowski@qt.io>