Fix visibility regression when using offscreen window in qquickwidget
Since s 7ded77d in declarative we keep offscreen window in QQuickWidget
as not visible to make input delivery system happy and due to fact that
window is really not visible. However, this breaks some behavior
in case that "emulated" visibility is required.

Switch to track qwindow in case of rendercontrol instead of internal
offscreen window.

Task-number: QTBUG-79213
Change-Id: If98deb8c56c3ae3c053f4e4d21e6b3b69211a9c3
Reviewed-by: Christian Strømme <christian.stromme@qt.io>