Implement QtAndroidWebViewController.getUserAgent
Commit 0a8a69d51042b95649dbdf471acb3c96228f372e introduced the
httpUserAgent property, with implementations for apple, winrt, webengine
and android.
However, the android implementation is missing a piece of
implementation.
Additionally, use semaphores to make sure that, on return, the setter
has completed setting the user agent, and therefore the emitted signal
is fired with the correct value.

Task-number: QTBUG-68746
Change-Id: Ic8f2a398297c8cd7d87ef4946ef26399167aeba9
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>
Reviewed-by: Christian Strømme <christian.stromme@qt.io>