Dynamically create the qrc file so it has the right alias
If the qrc file is created via the qml_module.prf then it takes
the path to the file as part of the alias and therefore this is in
the wrong place when it is loaded statically. So we need to create
the qrc file dynamically as well so that it correctly locates the
qmldir at the right location.

Change-Id: I620febf818a335319dafd6e5d53329b856369d68
Reviewed-by: Christian Strømme <christian.stromme@qt.io>