Do the pre-initialization steps after the plugin paths are set
Since initialize() is called before the creation of the application
then we need to delay the actual implementation until the paths are
set. Therefore a prehook is used for the application object so that
these can be done before the application finishes initalizing. For
static builds we call this function right away, thus requiring the
need for the QtWebView::initialize() function to still exist.

Change-Id: Ic4633cd1e96c39bc12f475fdd22c56646ea7ca10
Reviewed-by: Christian Strømme <christian.stromme@qt.io>
(cherry picked from commit 8e82991875c82916785823ba01a02614b04369de)