Sync with window visibilty due to offscreen windows in use
When there is an offscreen window used then this is always not visible,
but visiblityChanged is still emitted in this case and can be
relied upon. So we check if this is set to QWindow::Hidden and make
the webview visible/hidden accordingly.

Change-Id: I95892ef5df8b03feb282b49c8979ae6c93e18a87
Reviewed-by: Christian Strømme <christian.stromme@qt.io>
(cherry picked from commit c4cd14a0e8d199669554824ce917ad0db31373ce)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>