Update commercial license headers
Updated header.COMM to files in tqtc-qtwebview. Examples,
documentation or tests are not updated.

The commercial license header may contain some
additional lines so that its line count equals
with the earlier license header. Reason for this
is that some autotests use hard coded line numbers
and a change in the line count causes failures in
tests.

Task-number: QTQAINFRA-4208
Change-Id: Ib8f88a3637217a98e85116f01f2a6a8b4d8b8562
Reviewed-by: Jukka Jokiniva <jukka.jokiniva@qt.io>
Reviewed-by: Christian Strømme <christian.stromme@qt.io>