Fix assert when WebView is used in combination with QQuickWidget
Using the WebView with QQuickWidget is not really a use-case the
WebView is well suited for, due to the native WebView being an overlay,
however it should not crash or trigger an assert, so this change tries
to avoid some of the caveats when mixing with QQuickWidget. For example,
we cannot under any circumstances call winId() before the window the
QQuickWidget lives under, is backed by a platform window. The native
WebView is therefore not added before the platform window is created.

Pick-to: 6.4 6.2 5.15
Fixes: QTBUG-46084
Change-Id: I815d37cdd0328b3a258ef60294b5ea282f41cfc6
Reviewed-by: Christian Strømme <christian.stromme@qt.io>