Update commercial license headers
Updated header.COMM to the files in tqtc-qtwebview.
Examples, tests, or documentation files are not updated.

The commercial license header may contain some
additional lines so that its line count equals with
the earlier license header. Reason for this is that
some autotests use hard coded line numbers and a
change in the the line count causes failures in tests.

Task-number: QTQAINFRA-4954
Change-Id: I1e6fae709dd1ef32d35e9c8b89f6dc1f633c102c
Reviewed-by: Christian Strømme <christian.stromme@qt.io>