Android: Fix for possible thread deadlock
QtAndroidWebViewController constructor blocks a qGuiThread until
the WebView is created and configured in UI thread. Such an
implementation can easily lead to the deadlock (For example in the case
of handling InputConnection - it is exactly the opposite. UI thread is
blocked and waits for handling event by qGuiThread). That's why we need
to lock AndroidDeadlockProtector before blocking qGuiThread.

This implementation resolves potential deadlock using
AndroidDeadlockProtector. If protector cannot by locked, qGuiThread will
handle events to make it possible.

Fixes: QTBUG-82810
Pick-to: 6.5 6.4 6.2 5.15
Change-Id: I71aaea9ceb7c41b818ed533ce41c70b5c0e8d7de
Reviewed-by: Michal Klocek <michal.klocek@qt.io>