Android: make functions static if they don't need the object
This prevents threading issues when calling object functions in a lambda
run on the android main thread, when the QJniObject was instantiated on
the Qt main thread.

As a drive-by, store the id as a 64bit value on the C++ side. jlong is
always a 64bit type, even on 32bit Android systems, so we can't use
quintptr without truncating data.

Change-Id: I96d983eb9b2e99338768d30f7000ec9e75d3fea6
Reviewed-by: Christian Strømme <christian.stromme@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>