Android: modernize implementation
Declare used types, use modern variadic template APIs, and native
method declarations to get rid of all hand-crafted signature
strings.

In the native methods, cast the id (that was previously created
via reinterpret_cast from 'this', pointing to the private) back to
the private pointer directly, and only maintain a set of privates
to make sure that the id we get is still representing an alive
private.

Remove some dead (commented) code, and simplify JNI_OnLoad.

Change-Id: Ie8711e5f072edde5781b33941aee4e37b8213ea2
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>