CMake: Fix multi-arch iOS builds due to absolute framework paths
find_library(FWWebKit WebKit) stores the absolute path to the
framework of one of the sysroots, into FWWebKit.
When building with CMake 3.27 targeting multi-arch iOS, this results
in the following flag being passed to the compiler, regardless of
the architecture/sysroot:

  -F/Applications/Xcode14.app/Contents/Developer/Platforms/
  iPhoneOS.platform/Developer/SDKs/iPhoneOS16.2.sdk/
  System/Library/Frameworks

This causes warnings when building qdarwinwebview.mm regarding
unsupported webkit features during the x86_64 simulator build. When
combined with -Werror, it fails the build.

Use qt_internal_find_apple_system_framework() instead of
find_library(), to ensure we pass '-framework WebKit' instead of the
absolute path to the framework. This avoids warnings and is
cross-sysroot compatible.

As a drive-by, also fix the lookup of the Foundation framework. It's
called Foundation, not FoundationKit.

Pick-to: 6.5 6.2
Task-number: QTBUG-118138
Change-Id: I4379f14d65a37470f08ae8e7f46e00b886eadaf1
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>
(cherry picked from commit 7b29c1fe6fea9ee1a13082ba72483db76ce35719)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>