windows: Fix freeze when loading Qt Web Engine plugin
The mechanism which initializes Qt Web Engine via the plugin loader
does not work on some graphics drivers on Windows, which do not
support initializing graphics in DllMain(). Since Qt Web Engine is
always the backend on Windows, this patch makes Qt Web View depend
on it directly and calls initialize from QtWebView::initialize()
instead on that platform.

[ChangeLog][Windows] Fixed a freeze on startup on Windows. As part
of the solution for this, the Qt Web View library now depends
directly on Qt Web Engine on Windows, instead of indirectly via
the plugin library.

Fixes: QTBUG-117882
Change-Id: Ia1b3e54145477f645bbb97856bdbbb68b58d2785
Reviewed-by: Christian Strømme <christian.stromme@qt.io>