Add support for NSArray and NSDictionary in Darwin WebView's fromJSValue
To not have to write NSArray and NSDictionary conversion from scratch
we round-trip via JSON. One quirk of this approach is that NSDate is
supported as a standalone result object, but not nested in an array
or dictionary, as NSJSONSerialization doesn't support that. But at
least this brings us some of the way for array and dictionary support.

Fixes: QTBUG-70166
Pick-to: 6.5
Change-Id: Ied8a261e964c5adbadb503efb0bb7812ef911b83
Reviewed-by: Christian Strømme <christian.stromme@qt.io>
(cherry picked from commit b1e326ddffc45dd12dd6e454254649686bfdbfa4)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit 3c3f5813166a04f347f7083fe6a78251db85caac)