WebView2: Block creation of new windows not controlled by Qt
By default, an unhandled NewWindowRequested event will trigger
the creation of a window not controlled by us, and which can
outive the spawning process. This is inconsistent with other backends,
so handle the event and don't pass a window handle so the new window is
never created in the first place.

Fixes: QTBUG-139641
Change-Id: I75ed8ae376be1695f0f75c7286fed7dd3c7b3670
Reviewed-by: Michal Klocek <michal.klocek@qt.io>
(cherry picked from commit 5c29704a2ed0003edcf9ee7e591e4d392cce4d04)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>