Use QWebView as parent window in backends
Each backend besides QtWebEngine, provides
native window to be set by setContainedWindow in
QQuickWebView.

The android backend will have QWebView as parent
window, the same as darwin plugin.

Wasm and Webview2 do not create native child window.

Note the wasm in handled in follow up commit.

Note we do note make use of qquickwindow for webengine
in this commit, we barely just store the pointer,
or create the window if plugin was created without
the qquickwindow.

Change-Id: Ie98dc0103e353c280ebcb7add3c66f43d3618314
Reviewed-by: Moss Heim <moss.heim@qt.io>