WASM: Replace Quick-exclusive geometry hooks
The WASM backend relies on injecting an iframe into the generated
webpage, and then using Qt Quick ItemChange events to pass down
the correct geometry to the iframe's size. This meant that several
non-plugin classes would have ugly Q_OS_WASM ifdefs, and that the
WASM backend didn't work properly with C++.

This change removes all of the existing geometry change code,
and replaces it with a single JavaScript function that gets run when
the iframe is created. The JavaScript code installs a ResizeObserver
on the clientArea object (which represents the QWebView), and
changes the iframe's dimensions as needed.

Remove also obsolete now nativeWindowChanged().

Task-number: QTBUG-131837
Change-Id: I484a4b65406fa6b22ce969393d06e3c7b4dc3663
Reviewed-by: Moss Heim <moss.heim@qt.io>