Cleanup QQuickWebView initialization
In case of QQuickWebview with web engine backend,
we need to pass qt quick parent item. So far
it was was done by setting parent for QWebView
and then extracting it on web engine plugin backend's
initialization. Add initialize() call
instead. For time being use just QObject as parameter as it
is only used by webengine to pass parent QQuickItem.

Change-Id: I8d84313e93e2ad67ec1f008bee532f20b0d0a68e
Reviewed-by: Moss Heim <moss.heim@qt.io>
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>